backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
kafka-python==2.0.2
pydantic==1.8.2
python-multipart==0.0.5